feat(tabs): export old tabs as LegacyTabs
