+++ packages/bezier-react/src/components/Tabs/LegacyTabs/TabAction/TabAction.styled.ts
