+++ packages/bezier-react/src/components/Tabs/LegacyTabs/TabAction/index.ts
