+++ packages/bezier-react/src/components/Tabs/LegacyTabs/TabItem/TabItem.stories.tsx
@@ -5,7 +5,7 @@ import { Story, Meta } from '@storybook/react'
 
 /* Internal dependencies */
 import { getObjectFromEnum, getTitle } from 'Utils/storyUtils'
-import TabsSize from 'Components/Tabs/TabsSize'
+import TabsSize from 'Components/Tabs/LegacyTabs/TabsSize'
 import TabItem from './TabItem'
 import TabItemProps from './TabItem.types'
