+++ packages/bezier-react/src/components/Tabs/LegacyTabs/TabItem/TabItem.test.tsx
