+++ packages/bezier-react/src/components/Tabs/LegacyTabs/Tabs.types.ts
@@ -1,0 +1,8 @@
+/* Internal dependencies */
+import TabsSize from 'Components/Tabs/LegacyTabs/TabsSize'
+
+export default interface TabsOptions {
+  height?: TabsSize | number
+  withIndicator?: boolean
+  indicatorThickness?: number
+}
