+++ packages/bezier-react/src/components/Tabs/LegacyTabs/Tabs/Tabs.stories.tsx
@@ -7,9 +7,9 @@ import { Story, Meta } from '@storybook/react'
 
 /* Internal dependencies */
 import { getObjectFromEnum, getTitle } from 'Utils/storyUtils'
-import { TabItem } from 'Components/Tabs/TabItem'
-import { TabAction } from 'Components/Tabs/TabAction'
-import TabsSize from 'Components/Tabs/TabsSize'
+import { TabItem } from 'Components/Tabs/LegacyTabs/TabItem'
+import { TabAction } from 'Components/Tabs/LegacyTabs/TabAction'
+import TabsSize from 'Components/Tabs/LegacyTabs/TabsSize'
 import Tabs from './Tabs'
 import TabsProps from './Tabs.types'
