+++ packages/bezier-react/src/components/Tabs/LegacyTabs/Tabs/Tabs.styled.ts
