+++ packages/bezier-react/src/components/Tabs/LegacyTabs/Tabs/Tabs.tsx
@@ -4,8 +4,8 @@ import { v4 as uuid } from 'uuid'
 import { noop, isNumber, isEmpty, get, identity } from 'lodash-es'
 
 /* Internal dependencies */
-import { isTabItem } from 'Components/Tabs/TabItem'
-import { isTabAction } from 'Components/Tabs/TabAction'
+import { isTabItem } from 'Components/Tabs/LegacyTabs/TabItem'
+import { isTabAction } from 'Components/Tabs/LegacyTabs/TabAction'
 import TabsProps from './Tabs.types'
 import { Wrapper, TabItemWrapper, TabActions } from './Tabs.styled'
