+++ packages/bezier-react/src/components/Tabs/LegacyTabs/Tabs/index.ts
