+++ packages/bezier-react/src/components/Tabs/LegacyTabs/TabsSize.ts
