+++ packages/bezier-react/src/components/Tabs/LegacyTabs/index.ts
@@ -1,0 +1,3 @@
+export * from './TabAction'
+export * from './TabItem'
+export * from './Tabs'
