+++ packages/bezier-react/src/components/Tabs/TabList.styled.ts
