+++ packages/bezier-react/src/components/Tabs/Tabs.stories.tsx
