+++ packages/bezier-react/src/components/Tabs/Tabs.styled.ts
